refactored subdiags/tri, updated docstrings, subdiags psasing all tests
